third_party/blink/web_tests/platform/mac/external/wpt/html/canvas/element/manual/wide-gamut-canvas/canvas-display-p3-drawImage-video-expected.txt

This is a testharness.js-based test.
[FAIL] Rec2020-3FF000000, Context display-p3, ImageData display-p3, scaleImage=false
  assert_true: Actual pixel value 233,51,35,255 is approximately equal to 255,0,9,255. expected true got false
[FAIL] Rec2020-3FF000000, Context display-p3, ImageData display-p3, scaleImage=true
  assert_true: Actual pixel value 233,51,35,255 is approximately equal to 255,0,9,255. expected true got false
[FAIL] Rec2020-222000000, Context srgb, ImageData srgb, scaleImage=false
  assert_true: Actual pixel value 136,0,0,255 is approximately equal to 186,0,0,255. expected true got false
[FAIL] Rec2020-222000000, Context srgb, ImageData srgb, scaleImage=true
  assert_true: Actual pixel value 136,0,0,255 is approximately equal to 186,0,0,255. expected true got false
[FAIL] Rec2020-222000000, Context srgb, ImageData display-p3, scaleImage=false
  assert_true: Actual pixel value 124,22,13,255 is approximately equal to 170,34,23,255. expected true got false
[FAIL] Rec2020-222000000, Context srgb, ImageData display-p3, scaleImage=true
  assert_true: Actual pixel value 124,22,13,255 is approximately equal to 170,34,23,255. expected true got false
[FAIL] Rec2020-222000000, Context display-p3, ImageData srgb, scaleImage=false
  assert_true: Actual pixel value 136,0,0,255 is approximately equal to 186,0,0,255. expected true got false
[FAIL] Rec2020-222000000, Context display-p3, ImageData srgb, scaleImage=true
  assert_true: Actual pixel value 136,0,0,255 is approximately equal to 186,0,0,255. expected true got false
[FAIL] Rec2020-222000000, Context display-p3, ImageData display-p3, scaleImage=false
  assert_true: Actual pixel value 124,22,13,255 is approximately equal to 169,0,3,255. expected true got false
[FAIL] Rec2020-222000000, Context display-p3, ImageData display-p3, scaleImage=true
  assert_true: Actual pixel value 124,22,13,255 is approximately equal to 169,0,3,255. expected true got false
Harness: the test ran to completion.
